pyproject.toml
[project]
name = "bytebabel"
version = "0.2.0"
description = "ByteBabel — real-time speech transcription & translation"
requires-python = ">=3.10"
# IMPORTANT: use a Python with Tcl/Tk support (not uv's standalone bundle)
# macOS: brew install python@3.12 → uv venv --python /usr/local/bin/python3.12
# Linux: apt install python3-tk
# Windows: install from python.org (includes Tk by default)
dependencies = [
    "customtkinter>=5.2.2",
    "keyring>=25.7.0",
    "numpy>=2.2.6",
    "python-dotenv>=1.2.2",
    "soniox>=2.2.0",
    "sounddevice>=0.5.5",
    "Pillow>=10.0.0",
]

[project.scripts]
bytebabel = "bytebabel.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bytebabel"]

[dependency-groups]
dev = [
    "pyinstaller>=6.19.0",
    "hatchling>=1.27.0",
]
